Conflicts:
	geonode/settings.py

+++ geonode/geoserver/helpers.py
@@ -232,9 +232,17 @@ def cascading_delete(cat, layer_name):
             try:
                 store = get_store(cat, name, workspace=ws)
             except FailedRequestError:
-                logger.debug(
-                    'the store was not found in geoserver')
-                return
+                if ogc_server_settings.DATASTORE:
+                    try:
+                        store = get_store(cat, ogc_server_settings.DATASTORE, workspace=ws)
+                    except FailedRequestError:
+                        logger.debug(
+                            'the store was not found in geoserver')
+                        return
+                else:
+                    logger.debug(
+                        'the store was not found in geoserver')
+                    return
             if ws is None:
                 logger.debug(
                     'cascading delete was called on a layer where the workspace was not found')
@@ -269,15 +277,17 @@ def cascading_delete(cat, layer_name):
         for s in styles:
             if s is not None and s.name not in _default_style_names:
                 try:
-                    cat.delete(s, purge=True)
+                    cat.delete(s, purge='true')
                 except FailedRequestError as e:
                     # Trying to delete a shared style will fail
                     # We'll catch the exception and log it.
                     logger.debug(e)
 
         # Due to a possible bug of geoserver, we need this trick for now
+        # TODO: inspect the issue reported by this hack. Should be solved
+        #       with GS 2.7+
         try:
-            cat.delete(resource)  # This will fail
+            cat.delete(resource, recurse=True)  # This may fail
         except:
             cat.reload()  # this preservers the integrity of geoserver
 
@@ -289,12 +299,23 @@ def cascading_delete(cat, layer_name):
             # GeoGig repository.
             return
         else:
-            try:
-                if not store.get_resources():
-                    cat.delete(store, recurse=True)
-            except FailedRequestError as e:
-                # Catch the exception and log it.
-                logger.debug(e)
+            if store.resource_type == 'coverageStore':
+                try:
+                    logger.info(" - Going to purge the " + store.resource_type + " : " + store.href)
+                    cat.reset()  # this resets the coverage readers and unlocks the files
+                    cat.delete(store, purge='all', recurse=True)
+                    cat.reload()  # this preservers the integrity of geoserver
+                except FailedRequestError as e:
+                    # Trying to recursively purge a store may fail
+                    # We'll catch the exception and log it.
+                    logger.debug(e)
+            else:
+                try:
+                    if not store.get_resources():
+                        cat.delete(store, recurse=True)
+                except FailedRequestError as e:
+                    # Catch the exception and log it.
+                    logger.debug(e)
 
 
 def delete_from_postgis(resource_name):
@@ -304,19 +325,20 @@ def delete_from_postgis(resource_name):
     """
     import psycopg2
     db = ogc_server_settings.datastore_db
-    conn = psycopg2.connect(
-        "dbname='" +
-        db['NAME'] +
-        "' user='" +
-        db['USER'] +
-        "'  password='" +
-        db['PASSWORD'] +
-        "' port=" +
-        db['PORT'] +
-        " host='" +
-        db['HOST'] +
-        "'")
+    conn = None
     try:
+        conn = psycopg2.connect(
+            "dbname='" +
+            db['NAME'] +
+            "' user='" +
+            db['USER'] +
+            "'  password='" +
+            db['PASSWORD'] +
+            "' port=" +
+            db['PORT'] +
+            " host='" +
+            db['HOST'] +
+            "'")
         cur = conn.cursor()
         cur.execute("SELECT DropGeometryTable ('%s')" % resource_name)
         conn.commit()
@@ -326,7 +348,11 @@ def delete_from_postgis(resource_name):
             resource_name,
             str(e))
     finally:
-        conn.close()
+        try:
+            if conn:
+                conn.close()
+        except Exception as e:
+            logger.error("Error closing PostGIS conn %s:%s", resource_name, str(e))
 
 
 def gs_slurp(
@@ -441,14 +467,15 @@ def gs_slurp(
             set_attributes(layer, overwrite=True)
 
             # Fix metadata links if the ip has changed
-            if not created and settings.SITEURL not in layer.link_set.metadata()[0].url:
-                layer.link_set.metadata().delete()
-                layer.save()
-                metadata_links = []
-                for link in layer.link_set.metadata():
-                    metadata_links.append((link.mime, link.name, link.url))
-                resource.metadata_links = metadata_links
-                cat.save(resource)
+            if layer.link_set.metadata().count() > 0:
+                if not created and settings.SITEURL not in layer.link_set.metadata()[0].url:
+                    layer.link_set.metadata().delete()
+                    layer.save()
+                    metadata_links = []
+                    for link in layer.link_set.metadata():
+                        metadata_links.append((link.mime, link.name, link.url))
+                    resource.metadata_links = metadata_links
+                    cat.save(resource)
 
         except Exception as e:
             if ignore_errors:
@@ -769,7 +796,7 @@ def save_style(gs_style):
     style, created = Style.objects.get_or_create(name=gs_style.name)
     style.sld_title = gs_style.sld_title
     style.sld_body = gs_style.sld_body
-    style.sld_url = gs_style.body_href()
+    style.sld_url = gs_style.body_href
     style.save()
     return style
 
@@ -809,8 +836,9 @@ def get_attribute_statistics(layer_name, field):
 
 
 def get_wcs_record(instance, retry=True):
-    wcs = WebCoverageService(ogc_server_settings.public_url + 'wcs', '1.0.0')
+    wcs = WebCoverageService(ogc_server_settings.LOCATION + 'wcs', '1.0.0')
     key = instance.workspace + ':' + instance.name
+    logger.debug(wcs.contents)
     if key in wcs.contents:
         return wcs.contents[key]
     else:
@@ -936,24 +964,35 @@ def _create_db_featurestore(name, data, overwrite=False, charset="UTF-8", worksp
     """
     cat = gs_catalog
     dsname = ogc_server_settings.DATASTORE
+
+    ds_exists = False
     try:
         ds = get_store(cat, dsname, workspace=workspace)
-
+        ds_exists = True
     except FailedRequestError:
         ds = cat.create_datastore(dsname, workspace=workspace)
-        db = ogc_server_settings.datastore_db
-        db_engine = 'postgis' if \
-            'postgis' in db['ENGINE'] else db['ENGINE']
-        ds.connection_parameters.update(
-            host=db['HOST'],
-            port=db['PORT'],
-            database=db['NAME'],
-            user=db['USER'],
-            passwd=db['PASSWORD'],
-            dbtype=db_engine
-        )
-        cat.save(ds)
-        ds = get_store(cat, dsname, workspace=workspace)
+
+    db = ogc_server_settings.datastore_db
+    db_engine = 'postgis' if \
+        'postgis' in db['ENGINE'] else db['ENGINE']
+    ds.connection_parameters.update(
+        {'validate connections': 'true',
+         'max connections': '10',
+         'min connections': '1',
+         'fetch size': '1000',
+         'host': db['HOST'],
+         'port': db['PORT'],
+         'database': db['NAME'],
+         'user': db['USER'],
+         'passwd': db['PASSWORD'],
+         'dbtype': db_engine}
+    )
+
+    if ds_exists:
+        ds.save_method = "PUT"
+
+    cat.save(ds)
+    ds = get_store(cat, dsname, workspace=workspace)
 
     try:
         cat.add_data_to_store(ds, name, data,
@@ -1398,8 +1437,12 @@ def wps_execute_layer_attribute_statistics(layer_name, field):
                                'layer_name': 'geonode:%s' % layer_name,
                                'field': field
                                })
-
-    response = http_post(url, request, timeout=ogc_server_settings.TIMEOUT)
+    response = http_post(
+        url,
+        request,
+        timeout=ogc_server_settings.TIMEOUT,
+        username=ogc_server_settings.credentials.username,
+        password=ogc_server_settings.credentials.password)
 
     exml = etree.fromstring(response)
 
@@ -1420,16 +1463,24 @@ def wps_execute_layer_attribute_statistics(layer_name, field):
 
     result['unique_values'] = 'NA'
 
+    return result
+
     # TODO: find way of figuring out threshold better
-    if result['Count'] < 10000:
-        request = render_to_string('layers/wps_execute_gs_unique.xml', {
-                                   'layer_name': 'geonode:%s' % layer_name,
-                                   'field': field
-                                   })
+    # Looks incomplete what is the purpose if the nex lines?
 
-        response = http_post(url, request, timeout=ogc_server_settings.TIMEOUT)
+    # if result['Count'] < 10000:
+    #     request = render_to_string('layers/wps_execute_gs_unique.xml', {
+    #                                'layer_name': 'geonode:%s' % layer_name,
+    #                                'field': field
+    #                                })
 
-        exml = etree.fromstring(response)
+    #     response = http_post(
+    #     url,
+    #     request,
+    #     timeout=ogc_server_settings.TIMEOUT,
+    #     username=ogc_server_settings.credentials.username,
+    #     password=ogc_server_settings.credentials.password)
+    #     exml = etree.fromstring(response)
 
 
 def style_update(request, url):
